# swagger.yaml
swagger: "2.0"
info:
  description: "Hotel CRUD based on MVC using node.js and express. Basic implementation of user, hotel, room, room services and reservation."
  version: "1.0.0"
  title: "Hotel CRUD"
host: "127.0.0.1:3000/"
...
schemes:
- "http"
paths:
...
  /user/fetchReservations:
    get:
      tags:
      - "user"
      summary: "Find User Reservations"
      description: ""
      operationId: "fetchUserReservations"
      produces:
      - "application/json"
      parameters:
        - name: token
          in: header
          type: string
          required: true
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/FetchReservationsResponse"
        "400":
          description: "Invalid data"
...
  /reservation/:
    post:
      tags:
      - "reservation"
      summary: "User make a reservation of a room"
      description: ""
      operationId: "createReservation"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "starting_date"
        in: "body"
        description: "Starting date of reservation"
        required: false
        type: "string"
        format: "YYYY-MM-DD"
      - name: token
        in: header
        type: string
        required: true
      - name: "ending_date"
        in: "body"
        description: "Starting date of reservation"
        required: false
        type: "string"
        format: "YYYY-MM-DD"
      - name: "room_id"
        in: "body"
        description: "ID of the room"
        required: true
        type: integer
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/CreateReservationResponse"
    put:
      tags:
      - "reservation"
      summary: "Update starting/ending date of reservation"
      description: ""
      operationId: "updateReservation"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        type: "string"
        name: "starting_date"
        description: "Starting date to update"
        required: false
      - name: token
        in: header
        type: string
        required: true
      - in: "body"
        type: "string"
        name: "ending_date"
        description: "Ending date to update"
        required: false
      - in: "body"
        type: "integer"
        name: "room_id"
        description: "ID of the room for which updating the reservation"
        required: true
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "Invalid Order"
...
  /room/:
    post:
      tags:
      - "room"
      summary: "Create a room in hotel"
      description: ""
      operationId: "createRoom"
      produces:
      - "application/json"
      parameters:
      - name: "price"
        in: "body"
        description: "Price of the room"
        required: true
        type: "integer"
      - name: hotel_id
        in: body
        description: "ID of the hotel to create a room in it"
        type: integer
        required: true
      - name: token
        in: header
        type: string
        required: true
      responses:
        "200":
          description: "Successfully Added"
    put:
      tags:
      - "room"
      summary: "Update a room in hotel"
      description: ""
      operationId: "updateRoom"
      produces:
      - "application/json"
      parameters:
      - name: "price"
        in: "body"
        description: "Price of the room"
        required: false
        type: "integer"
      - name: token
        in: header
        type: string
        required: true
      - name: room_id
        in: body
        description: "ID of the room to be updated"
        type: integer
        required: true
      responses:
        "200":
          description: "Successfully updated"
          schema:
            $ref: "#/definitions/CreateRoomResponse"
    get:
      tags:
      - "room"
      summary: "Fetch rooms of a hotel"
      description: ""
      operationId: "fetchRooms"
      produces:
      - "application/json"
      parameters:
      - name: "hotel_id"
        in: "body"
        description: "ID of the hotel to get its rooms"
        required: true
        type: "integer"
      - name: token
        in: header
        type: string
        required: true
      responses:
        "200":
          description: "List of rooms in a hotel"
          schema:
            $ref: "#/definitions/FetchRoomResponse"
definitions:
  CreateReservationResponse:
    type: object
    properties:
        id:
          type: integer
          example: 1
        starting_date:
          type: string
          example: "2021-12-01T00:00:00.000Z"
        ending_date:
          type: string
          example: "2021-12-01T00:00:00.000Z"
        customer_id:
          type: integer
          example: 3
        createdAt:
          type: integer
          example: "2021-12-01T00:00:00.000Z"
        updatedAt:
          type: integer
          example: "2021-12-01T00:00:00.000Z"
        room_id:
          type: integer
          example: 2
...
  FetchReservationsResponse:
    type: "array"
    items:
      $ref: '#/definitions/CreateReservationResponse'
...
  FetchRoomResponse:
    type: "array"
    items:
      $ref: '#/definitions/CreateRoomResponse'
...
  CreateRoomResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      status:
        type: "string"
      price:
        type: "integer"
      hotel_id:
        type: integer
      createdAt:
        type: "string"
      updatedAt:
        type: "string"